removed expand script and merged functionality
